requirements.txt
flask
flask-sqlalchemy
pytest
Flask-WTF
flask-login
wtforms
bs4
python-dotenv
Flask-Migrate